requirements.txt
#   pip install django-storages[sftp]

aioredis~=2.0.1
asgiref~=3.11.0
attrs~=25.4.0
Authlib~=1.6.5
celery~=5.6.3
certifi~=2026.2.25
cffi~=2.0.0
channels~=4.3.0
channels-redis~=4.3.0
charset-normalizer~=3.4.4
colorama~=0.4.6
coverage~=7.12.0
cryptography~=46.0.3
daphne~=4.2.1
decorator~=5.2.1
Django~=6.0.3
django-auto-prefetch~=1.14.0
django_celery_results~=2.6.0
django-colorfield~=0.14.0
django-cors-headers~=4.9.0
django-dbbackup~=5.2.0
django-debug-toolbar~=6.2.0
django-ipware~=7.0.1
django-picklefield~=3.4.0
django-polymorphic~=4.11.2
django-redis~=6.0.0
django-rest-auth~=0.9.5
django-rest-polymorphic~=0.1.10
###django-silk~=5.1.0
djangorestframework~=3.16.1
djangorestframework-oauth~=1.1.0
djangorestframework-simplejwt~=5.5.1
docker~=7.1.0
eventlet~=0.40.4
factory-boy~=3.3.3
Faker~=38.2.0
fuzzywuzzy~=0.18.0
gunicorn~=25.3.0
hiredis~=3.3.1
html5lib~=1.1
jsonfield~=3.2.0
kiwisolver~=1.4.9
kombu~=5.6.1
Levenshtein~=0.27.3
license-expression~=30.4.4
matplotlib~=3.10.7
mdurl~=0.1.2
msgpack~=1.1.2
numpy==2.2.6
odfpy~=1.4.1
#opencv-python-headless==4.12.0.88
openpyxl~=3.1.5
packaging~=25.0
pandas~=2.3.3
Pillow~=12.0.0
pip-requirements-parser~=32.0.1
pip-review~=1.3.0       # pip-review --interactive      ||  pip-review -aC
pip-upgrader~=1.4.15
pip_audit~=2.9.0
plotly~=6.5.0
psycopg2-binary~=2.9.11
py-serializable~=2.1.0
pydantic~=2.12.3
pydantic_core~=2.41.4
PyJWT~=2.10.1
pyOpenSSL~=25.3.0
pyparsing~=3.2.5
pypdf~=6.4.0
pypng~=0.20220715.0
python-dateutil~=2.9.0.post0
python-gnupg~=0.5.5
python-Levenshtein~=0.27.3
python-memcached~=1.62
pytz~=2025.2
pytest~=9.0.0
pytest-cov~=7.0.0
pytest-django~=4.11.1
#pytest-playwright~=0.5.2
pyzipper~=0.3.6
qrcode~=8.2
RapidFuzz~=3.14.1
redis~=7.1.0
rest-pandas~=1.1.0
requests~=2.32.5
requests_cache~=1.2.1
requests_ratelimiter~=0.7.0
resolvelib~=1.2.1
rncryptor~=3.3.0
ruff~=0.15.8            # ruff check --fix
simplejson~=3.20.2
tablib~=3.9.0
typing_extensions~=4.15.0
urllib3~=2.5.0
uvicorn~=0.38.0
websockets~=15.0.1
XlsxWriter~=3.2.9


# ================
# === D E V ======
# ================

# safety~=3.2.13
# safety-schemas~=0.0.8
# pip install safety pip-audit
# pip-audit
# safety scan

# pip freeze > requirements.txt
# factory_boy~=3.3.3
# playwright~=1.55.0
